refactor(tor/hs-host): cleanup — shared crypto in tor-crypto, leak fixes, race guard
Cleanup pass on the hidden-service host based on a code review of 967bd9f:

- Lifted the rend-spec-v3 crypto helpers (sha3, mac, dMac, kdfShake256,
  u64be, bytesToBigIntLE/bigIntToBytesLE, modInverse, Sha3_256Hash,
  base32EncodeLowerNoPad) out of `packages/tor/src/hidden-service.ts`
  and `packages/tor/src/hidden-service-host.ts` into the shared
  `tor-crypto` package as `packages/crypto/src/hs-crypto.ts`. Re-exported
  through both `node.ts` and `browser.ts` entry points so a single
  isomorphic implementation serves the client + the server, on Node and
  in the browser.

  Net result: removes ~200 LoC of crypto duplication; both consumers
  now import the same primitives from the same place.

- Fixed two real resource leaks on error paths:
  * `uploadDescriptor()` — the HSDir circuit was only destroyed on the
    success branch; an exception between buildCircuitToTarget() and the
    final destroy() leaked it. Wrapped in try/finally.
  * `handleIntroduce2()` — once the rendezvous circuit was built, any
    failure before the `'rendezvous'` event was emitted leaked a 3-hop
    circuit per failed handshake. Now destroyed unless we hand it off
    to the application.

- Detached intro / rendezvous circuit `'relay'` listeners on circuit
  `'destroyed'`. Without this, the closures pinned `this` and `intro`
  for the lifetime of the host — a real leak under intro-point churn.

- Guarded against concurrent `start()` calls by adding a synchronous
  `starting` latch (cleared in finally). Two parallel start()s no longer
  both pass the `if (this.running)` check and end up establishing
  duplicate intro points.

- Dropped the `stop()` synchronous alias; `unpublish()` is the documented
  API and there are no in-tree callers of `stop()` to migrate.

Verified via tsc --noEmit (clean modulo pre-existing missing optional
deps) and a 24-step in-process smoke harness covering blinded EdDSA,
ESTABLISH_INTRO byte format + signature, INTRODUCE2 round-trip vs the
client implementation, hs-ntor server↔client cipher-pair agreement, and
descriptor structure.
